do not remove Staging jobs from TaskQueue
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@10790 4525493e-7705-40b1-a816-d608a930855b

+++ WorkloadManagementSystem/Service/MatcherHandler.py
@@ -1,5 +1,5 @@
 ########################################################################
-# $Id: MatcherHandler.py,v 1.14 2008/07/18 08:33:27 rgracian Exp $
+# $Id: MatcherHandler.py,v 1.15 2008/07/18 15:36:29 rgracian Exp $
 ########################################################################
 """
 Matcher class. It matches Agent Site capabilities to job requirements.
@@ -7,7 +7,7 @@ It also provides an XMLRPC interface to the Matcher
 
 """
 
-__RCSID__ = "$Id: MatcherHandler.py,v 1.14 2008/07/18 08:33:27 rgracian Exp $"
+__RCSID__ = "$Id: MatcherHandler.py,v 1.15 2008/07/18 15:36:29 rgracian Exp $"
 
 import re, os, sys, time
 import string
@@ -261,9 +261,10 @@ class MatcherHandler(RequestHandler):
               gLogger.debug("Job JDL:\n"+classAdJob.asJDL())
           else:
             gLogger.warn("Job %d in the Task Queue but the status is %s" % (job,status))
-            result = jobDB.deleteJobFromQueue(job)
-            if not result['OK']:
-              gLogger.warn("Failed to delete job %d from Task Queue" % job)
+            if status != 'Staging':
+              result = jobDB.deleteJobFromQueue(job)
+              if not result['OK']:
+                gLogger.warn("Failed to delete job %d from Task Queue" % job)
         else:
           gLogger.warn("Job %d ot found in the JobDB, will be deleted from the Task Queue" % job)
           result = jobDB.deleteJobFromQueue(job)
